code review fixes
